% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_model.R
\name{build_rf}
\alias{build_rf}
\alias{build_xgb}
\alias{build_brnn}
\title{Build RT prediction model}
\usage{
build_rf(
  trainingDf,
  k = 10,
  percentage = 0.8,
  seed = 1,
  search = "random",
  grid = NULL,
  metric = "Rsquared",
  thread = 1
)

build_xgb(
  trainingDf,
  k = 10,
  percentage = 0.8,
  seed = 1,
  search = "random",
  grid = NULL,
  metric = "Rsquared",
  thread = 1
)

build_brnn(
  trainingDf,
  k = 10,
  percentage = 0.8,
  seed = 1,
  search = "random",
  grid = NULL,
  metric = "Rsquared",
  thread = 1
)
}
\arguments{
\item{trainingDf}{A training data frame.}

\item{k}{Either the number of folds or number of resampling iterations.}

\item{percentage}{For leave-group out cross-validation: the training percentage.}

\item{seed}{Seed number.}

\item{search}{Either "grid" or "random", describing how the tuning parameter grid is determined.}

\item{grid}{A data frame with possible tuning values.}

\item{metric}{A string that specifies what summary metric will be used to select the optimal model.}

\item{thread}{Number of threads in parallel.}
}
\value{
A model list.
}
\description{
Build a machine learning model.
}
\details{
These functions use the caret packages.
For more information about parameters, please refer to \code{\link[caret]{train}} and \code{\link[caret]{trainControl}} .
}
\examples{
model_rf <- build_rf(trainingDf = trainingDf)
model_xgb <- build_xgb(trainingDf = trainingDf)
model_brnn <- build_brnn(trainingDf = trainingDf)
}
