% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMARTS.R
\name{C_C_count}
\alias{C_C_count}
\title{C_C_count}
\usage{
C_C_count(smiles, start_atom_idx, end_atom_idx)
}
\arguments{
\item{smiles}{A smiles string.}

\item{start_atom_idx}{A number.}

\item{end_atom_idx}{A number.}
}
\value{
Atom number.
}
\description{
Atom number between two C atom.
}
\examples{
draw_smiles(smiles = "C(O)(=O)CC(O)C/C=C\\\\C/C=C\\\\C/C=C\\\\C/C=C\\\\CCCCC", SMARTS = "[CX3;$(C=O);$(C-O)](=O)O")
smartsMatch(smiles = "C(O)(=O)CC(O)C/C=C\\\\C/C=C\\\\C/C=C\\\\C/C=C\\\\CCCCC", SMARTS = "[CX3;$(C=O);$(C-O)](=O)O")
C_C_count("C(O)(=O)CC(O)C/C=C\\\\C/C=C\\\\C/C=C\\\\C/C=C\\\\CCCCC", start_atom_idx = 0, end_atom_idx = 7)
}
