% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_smiles.R
\name{draw_smi}
\alias{draw_smi}
\title{Draw a molecule using a SMILES}
\usage{
draw_smi(
  smi,
  width = 450,
  height = 150,
  addAtomIndices = TRUE,
  addBondIndices = FALSE,
  addStereoAnnotation = FALSE,
  explicitMethyl = FALSE
)
}
\arguments{
\item{smi}{A SMILES string}

\item{width}{Width of pic}

\item{height}{Height of pic}

\item{addAtomIndices}{Whether to add atom indices}

\item{addBondIndices}{Whether to add bond indices}

\item{addStereoAnnotation}{Whether to add stereoisomorphic annotations}

\item{explicitMethyl}{Whether to display the terminal methyl group}
}
\description{
Draw a molecule using a SMILES
}
\examples{
draw_smi(smi = "C(OCCCCCCCC(CC)(CCC)CCCCC(C)C)C(OC(CCCCCCCCCCCC(C)C)=O)COC(=O)CCCCCCC(OCCC)CCCCC(C)C",
         width = 1600, height = 800)
}
