% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_smiles.R
\name{draw_smiles}
\alias{draw_smiles}
\title{Draw molecules using smiles}
\usage{
draw_smiles(smiles)
}
\arguments{
\item{smiles}{A vector with smiles.}
}
\value{
None.
}
\description{
Draw molecules using smiles
}
\examples{
draw_smiles(smiles = c("c1ccccc1O", "CCCCC"))
}
