% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{predictRt}
\alias{predictRt}
\alias{evaluate_model}
\title{Evaluate model}
\usage{
predictRt(testingDf, model)

evaluate_model(testingDf, model, digits = 3)
}
\arguments{
\item{testingDf}{A testing data frame.}

\item{model}{Model list.}

\item{digits}{integer indicating the number of decimal places (round) or significant digits (signif) to be used}
}
\description{
Evaluate model.
}
\examples{
predictRt(testingDf = testingDf, model = model_rf)
evaluate_model(testingDf = testingDf, model = model_rf)
}
