% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3D.R
\name{get_atom_distance}
\alias{get_atom_distance}
\title{get_atom_distance}
\usage{
get_atom_distance(smiles, start_atom_idx, end_atom_idx)
}
\arguments{
\item{smiles}{A smiles string.}

\item{start_atom_idx}{Start atom index.}

\item{end_atom_idx}{End atom index.}
}
\value{
A distance value.
}
\description{
Calculate 3D distance between two atom.
}
\examples{
smiles <- "C(OCC)(/C=C/C(O)CCCCC)C(O)CCCCCCCC(=O)O"
get_atom_distance(smiles, start_atom_idx = 6, end_atom_idx = 22)
get_atom_distance(smiles, start_atom_idx = 13, end_atom_idx = 22)
}
