% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMD.R
\name{GetCDK_MD}
\alias{GetCDK_MD}
\title{GetCDK_MD}
\usage{
GetCDK_MD(cmpDf, flavor = "CxSmiles", category = "all", thread = 1)
}
\arguments{
\item{cmpDf}{A tibble or data.frame with two column, id and smiles.}

\item{flavor}{SMILES flavor.}

\item{category}{cdk molecule descriptors category.}

\item{thread}{Parallel thread.}
}
\value{
A tibble.
}
\description{
Calculate cdk molecule descriptors for a cmpDf.
}
\examples{
data("cmpDf_demo", package = "LipRtPred")
descsDf <- GetCDK_MD(cmpDf = cmpDf_demo)
}
