% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMD.R
\name{GetMordred_MD}
\alias{GetMordred_MD}
\title{GetMordred_MD}
\usage{
GetMordred_MD(cmpDf, flavor = "CxSmiles", ignore_3D = TRUE, thread = 1)
}
\arguments{
\item{cmpDf}{A tibble or data.frame with two column, id and smiles.}

\item{flavor}{SMILES flavor.}

\item{ignore_3D}{Whether to ignore 3D descriptors.}

\item{thread}{Parallel thread.}
}
\value{
A tibble.
}
\description{
Calculate mordred molecule descriptors for a cmpDf.
}
\examples{
data("cmpDf_demo", package = "LipRtPred")
descsDf <- GetMordred_MD(cmpDf = cmpDf_demo)
}
