% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMD.R
\name{GetRDKit_MD}
\alias{GetRDKit_MD}
\title{GetRDKit_MD}
\usage{
GetRDKit_MD(cmpDf, flavor = "CxSmiles", thread = 1)
}
\arguments{
\item{cmpDf}{A tibble or data.frame with two column, id and smiles.}

\item{flavor}{SMILES flavor.}

\item{thread}{Parallel thread.}
}
\value{
A tibble.
}
\description{
Calculate rdkit molecule descriptors for a cmpDf.
}
\examples{
data("cmpDf_demo", package = "LipRtPred")
descsDf <- GetRDKit_MD(cmpDf = cmpDf_demo)
}
