% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel.R
\name{parallelEnv}
\alias{parallelEnv}
\title{Generate parallel environment}
\usage{
parallelEnv(thread = parallel::detectCores())
}
\arguments{
\item{thread}{Number of parallel threads}
}
\value{
cl
}
\description{
Generate parallel environment
}
\examples{
parallelEnv()
}
