% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_data.R
\name{filterColumns}
\alias{filterColumns}
\alias{split_data}
\title{Preprocess raw data}
\usage{
filterColumns(
  inputDf,
  freqCut = 95/5,
  uniqueCut = 10,
  allowParallel = TRUE,
  cutoff = 0.9
)

split_data(inputDf, percentage = 0.8, seed = 1)
}
\arguments{
\item{inputDf}{Input data frame with id, smiles and MD or FP.}

\item{freqCut}{The cutoff for the ratio of the most common value to the second most common value.}

\item{uniqueCut}{The cutoff for the percentage of distinct values out of the number of total samples.}

\item{allowParallel}{should the parallel processing via the foreach package be used for the computations? If TRUE, more memory will be used but execution time should be shorter.}

\item{cutoff}{A numeric value for the pair-wise absolute correlation cutoff.}

\item{percentage}{The percentage of data that goes to training.}

\item{seed}{The seed number.}
}
\value{
A tibble.

A list with training and testing sets.
}
\description{
Preprocess raw data as follows: (1) remove columns containing NA; (2) remove columns with variance close to 0.
The data were then divided into training and testing data.
}
\details{
This function uses the caret package, for more information on \code{\link[caret]{nearZeroVar}}.
}
\examples{
data("cmpDf_demo", package = "LipRtPred")
descsDf <- GetCDK_MD(cmpDf = cmpDf_demo)
inputDf <- filterColumns(inputDf = descsDf)
tmp <- split_data(inputDf = inputDf, percentage = 0.8)
trainingDf <- tmp$training
testingDf <- tmp$testing
}
