% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMARTS.R
\name{smartsMatch}
\alias{smartsMatch}
\title{Match substructure using SMARTS}
\usage{
smartsMatch(smiles, SMARTS)
}
\arguments{
\item{smiles}{smiles vector.}

\item{SMARTS}{SMARTS pattern.}
}
\value{
list.
}
\description{
Match substructure using SMARTS
}
\examples{
smartsMatch(smiles = c("CC(O)CC(O)C", "C(O)CC(O)C"), SMARTS = "CO")
}
