% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMD.R
\name{GetCDK_MD}
\alias{GetCDK_MD}
\alias{GetRDKit_MD}
\alias{GetMordred_MD}
\title{Calculate molecule descriptors.}
\usage{
GetCDK_MD(cmpDf, flavor = "CxSmiles", category = "all", thread = 1)

GetRDKit_MD(cmpDf, flavor = "CxSmiles", thread = 1)

GetMordred_MD(cmpDf, flavor = "CxSmiles", ignore_3D = TRUE, thread = 1)
}
\arguments{
\item{cmpDf}{A tibble with two column, id and smiles.}

\item{flavor}{Default is CxSmiles. More flavor please see \code{\link[rcdk]{smiles.flavors}}.}

\item{category}{CDK molecule descriptors category include "all", "protein", "hybrid", "constitutional", "topological", "electronic" and "geometrical"}

\item{thread}{Number of threads in parallel.}

\item{ignore_3D}{Whether to ignore 3D descriptors.}
}
\value{
A tibble.
}
\description{
Three molecule descriptors can be computed, CDK, RDKit and Mordred.
}
\examples{
data("cmpDf_demo", package = "LipRtPred")
descsDf <- GetCDK_MD(cmpDf = cmpDf_demo)
descsDf <- GetRDKit_MD(cmpDf = cmpDf_demo)
descsDf <- GetMordred_MD(cmpDf = cmpDf_demo)
}
